Splitted the eurobraille display driver code into seperate files. Fixed some linting due to splitting.
