+++ source/brailleDisplayDrivers/eurobraille/__init__.py
@@ -1,0 +1,9 @@
+# A part of NonVisual Desktop Access (NVDA)
+# This file is covered by the GNU General Public License.
+# See the file COPYING for more details.
+# Copyright (C) 2023 NV Access Limited, Babbage B.V.
+
+"""Braille display driver for eurobraille displays."""
+
+# Imported here so that braille._getDisplayDriver can import
+from .driver import BrailleDisplayDriver  # noqa: F401
